// petconnect-frontend/src/types/petData.ts
export interface PetType {
  _id: string
  name: string
  description?: string
  icon: string
  isActive: boolean
  createdAt: string
  updatedAt: string
}

export interface Temperament {
  _id: string
  name: string
  description?: string
  icon: string
  isActive: boolean
  createdAt: string
  updatedAt: string
}